Do not use sdnotify by default
Notify is excellent as per https://www.redhat.com/sysadmin/podman-auto-updates-rollbacks
since it enables autoupdates,
but sometimes containers may fail to start because of this option.
Needs investigation.

Leave it present so users can search about it and test it themselves,
but be conservative by default (even if I'm not).