Remove net suffix from networks
At some point, I added this either because of some very old Podman limitation
or because it is easier to differentiate between the original container file
and the network file when they are not parallel in a text editor.
I know I did this originally for some reason in production years ago.

Either way it should be fine to not use the net suffix since the generated
network comes in the format `systemd-servicename` and the generated systemd
service comes in the format `servicename-network.service`.